Huge commit. Adds support for a notch.
Couple of notes. This API was somewhat difficult to create, but I think
I finally hit the right balance. In order to use a notch, you must
provide both a content view and a content background color. The
background color is so that I can draw the image/color down into the
notch, while still maintaining content aspect integrity.

The all-in-one creation method has been deprecated - properties are the
way to go.

Also, a bundle was added with static images for the shadows. This is
really the best and most optimized way to render these, as CG functions
just don't provide the same level of performance that images
(especially once cached) can provide.